backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
supabase==2.4.6
openai==1.30.1
requests==2.31.0
pandas==2.2.2
httpx==0.27.0
